Initial checkin test_sparse_ndarray passes
